refactor: moved utility logic to index
